dxf: WIP is_sorted_POLYLINE shortcut (#83)
COMMON_ENTITY_POLYLINE: ensure that the common POLYLINE_*
fields are shared with all 4 types.

Some vertex[] handles are even worse than a wrong SEQEND,
check when the first vertex[0] is corrupt and ignore it then.

enable loglevel LOG in out_dxf, mainly to warn about this
wrong POLYLINE.vertex[0] handle %lX < %lX

+++ src/dec_macros.h
@@ -633,7 +633,7 @@
     vcount = bit_position(dat); \
     if (obj->hdlpos != (unsigned long)vcount) { \
       bit_set_position(hdl_dat, obj->hdlpos); \
-      LOG_HANDLE(" handle stream: %+ld @%lu.%u %s\n", obj->hdlpos - vcount, \
+      LOG_HANDLE(" handle stream: %+ld @%lu.%u %s\n", (long)obj->hdlpos - (long)vcount, \
         dat->byte, dat->bit, \
         ((long)obj->hdlpos - (long)vcount) >= 8 ? "MISSING" \
           : ((long)obj->hdlpos < (long)vcount) ? "OVERSHOOT" : ""); \
